Move canDeclProvideDefaultImplementationFor() from libIDE into libIndex
